preductos criticos

+++ Sisfarma.Sincronizador.Domain.Core/ExternalServices/Fisiotes/IFaltasExternalService.cs
@@ -4,9 +4,12 @@ namespace Sisfarma.Sincronizador.Domain.Core.ExternalServices.Fisiotes
 {
     public interface IFaltasExternalService : IFaltasExternalServiceNew
     {
-        bool ExistsLineaDePedido(int idPedido, int idLinea);
-        Falta GetByLineaDePedido(int pedido, int linea);
+        bool ExistsLineaDePedido(long idPedido, int idLinea);
+
+        Falta GetByLineaDePedido(long pedido, int linea);
+
         void Insert(Falta ff);
+
         Falta LastOrDefault();
     }
